Remove Pop3Store.createUri/decodeUri
