+++ app/k9mail/src/main/java/com/fsck/k9/backends/Pop3BackendFactory.kt
@@ -10,6 +10,8 @@ import com.fsck.k9.mail.ServerSettings
 import com.fsck.k9.mail.oauth.OAuth2TokenProvider
 import com.fsck.k9.mail.ssl.DefaultTrustedSocketFactory
 import com.fsck.k9.mail.store.pop3.Pop3Store
+import com.fsck.k9.mail.store.pop3.Pop3StoreUriCreator
+import com.fsck.k9.mail.store.pop3.Pop3StoreUriDecoder
 import com.fsck.k9.mail.transport.smtp.SmtpTransport
 import com.fsck.k9.mail.transport.smtp.SmtpTransportUriCreator
 import com.fsck.k9.mail.transport.smtp.SmtpTransportUriDecoder
@@ -36,11 +38,11 @@ class Pop3BackendFactory(private val context: Context, private val preferences: 
     }
 
     override fun decodeStoreUri(storeUri: String): ServerSettings {
-        return Pop3Store.decodeUri(storeUri)
+        return Pop3StoreUriDecoder.decode(storeUri)
     }
 
     override fun createStoreUri(serverSettings: ServerSettings): String {
-        return Pop3Store.createUri(serverSettings)
+        return Pop3StoreUriCreator.create(serverSettings)
     }
 
     override fun decodeTransportUri(transportUri: String): ServerSettings {
